requirements.txt
environs>=9.5
gitpython>=3.1
json5>=0.9
Pillow>=9.2
toolz>=0.12

# Development
black>=22.6
mypy>=0.961
